telegram_rest_mvc/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-rest-mvc"
version = "0.1.0"
description = "Minimal MVC routing framework for python-telegram-bot >=20"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name="Your Name", email="you@example.com" }
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-telegram-bot>=22,<23",
]

[project.urls]
"Homepage" = "https://github.com/yourname/telegram-rest-mvc"
"Bug Tracker" = "https://github.com/yourname/telegram-rest-mvc/issues"

[tool.setuptools]
package-dir = {"telegram_rest_mvc" = ""}
